sample-Python-KeyVault-Function/requirements.txt
adal==1.2.1
asn1crypto==0.24.0
astroid==2.1.0
azure-common==1.1.17
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-keyvault==1.1.0
azure-nspkg==3.0.2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
cryptography==2.5
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
isodate==0.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
msrest==0.6.4
msrestazure==0.6.0
oauthlib==3.0.1
protobuf==3.6.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
python-dateutil==2.7.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1